pyproject.toml
[project]
name = "nonebot_plugin_sparkapi"
version = "1.5.0"
description = "Nonebot2 ChatBot that calls the official API of the iflyrec Spark LLM for all model versions, defaulting to the current latest (v4.0)"
authors = [
    { name = "CCLMSY", email = "2502408581@qq.com" }
]
dependencies = [
    "nonebot-adapter-onebot >=2.2.1, <3.0.0",
    "nonebot2 >=2.0.0rc3, <3.0.0",
    "websockets >=12.0, <13.0",
    "httpx >=0.19.0, <1.0.0",
    "pillow >=10.0.0, <11.0.0",
]
requires-python = ">=3.10"
license = { file = "LICENSE" }
readme = { file = "README.md", content-type = "text/markdown" }

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

